app/requirements-dev.txt
black
docstr-coverage
flask-debugtoolbar
ipdb
isort
pytest
pytest-cov
pytest-flask
pytest-repeat
